Getting rid of OrderConfirmVC
1. MenuForClientVC now handles everything in one page
2. Added some prettiness
3. Upon placing an order, init of OrderVM filters out the orderitemvms
that have qty<=0
